fixed script and added lazygit

+++ files/scripts/font_rendering_symlinks.sh
@@ -1,5 +1,13 @@
 #!/usr/bin/env bash
 set -oue pipefail
 
-ln -s /usr/share/fontconfig/conf.avail/11-lcdfilter-default.conf /etc/fonts/conf.d
-ln -s /usr/share/fontconfig/conf.avail/10-sub-pixel-rgb.conf /etc/fonts/conf.d
+CONF_DIR="/etc/fonts/conf.d"
+
+# if the symlink below isn't there then symlink from /usr/share
+if [ ! -L "$CONF_DIR/11-lcdfilter-default.conf" ]; then
+  ln -s /usr/share/fontconfig/conf.avail/11-lcdfilter-default.conf /etc/fonts/conf.d
+fi
+
+if [ ! -L "$CONF_DIR/10-sub-pixel-rgb.conf" ]; then
+  ln -s /usr/share/fontconfig/conf.avail/10-sub-pixel-rgb.conf /etc/fonts/conf.d
+fi
